remove dependency on mro
mro is only used in tests on perl versions that include it in core.
Listing it as a prereq makes it impossible to install this module on
older perl versions where it wouldn't be used anyway. It would be
possible to conditionally list it as a prereq, but since mro is core
only, this serves little purpose and is extra complexity.